Add TcpReader and TcpWriter, move IO methods to traits.
