+++ hl/src/lib.rs
@@ -97,14 +97,17 @@
 #![cfg_attr(docsrs, feature(doc_cfg))]
 #![cfg_attr(all(not(feature = "std"), not(test)), no_std)]
 
+mod tcp;
+mod udp;
+
 #[cfg(feature = "defmt")]
 use dfmt as defmt;
 
+use ll::{Registers, Sn, SocketCommand, SocketStatus, SOCKETS};
+
+pub use tcp::{Tcp, TcpReader, TcpWriter};
+pub use udp::{Udp, UdpHeader, UdpReader, UdpWriter};
 pub use w5500_ll as ll;
-
-use ll::{Protocol, Registers, Sn, SocketCommand, SocketMode, SocketStatus, SOCKETS};
-
-use core::cmp::min;
 
 /// Networking data types.
 ///
@@ -114,40 +117,6 @@ pub mod net {
 }
 
 use net::{Ipv4Addr, SocketAddrV4};
-
-/// W5500 UDP Header.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct UdpHeader {
-    /// Origin IP address and port.
-    pub origin: SocketAddrV4,
-    /// Length of the UDP packet in bytes.
-    ///
-    /// This may not be equal to the length of the data in the socket buffer if
-    /// the UDP packet was truncated.
-    pub len: u16,
-}
-
-impl UdpHeader {
-    // note: not your standard UDP datagram header
-    // For a UDP socket the W5500 UDP header contains:
-    // * 4 bytes origin IP
-    // * 2 bytes origin port
-    // * 2 bytes size
-    const LEN: u16 = 8;
-    const LEN_USIZE: usize = Self::LEN as usize;
-
-    /// Deserialize a UDP header.
-    fn deser(buf: [u8; Self::LEN_USIZE]) -> UdpHeader {
-        UdpHeader {
-            origin: SocketAddrV4::new(
-                Ipv4Addr::new(buf[0], buf[1], buf[2], buf[3]),
-                u16::from_be_bytes([buf[4], buf[5]]),
-            ),
-            len: u16::from_be_bytes([buf[6], buf[7]]),
-        }
-    }
-}
 
 fn port_is_unique<T: ?Sized, E>(w5500: &mut T, socket: Sn, port: u16) -> Result<bool, E>
 where
@@ -243,6 +212,13 @@ pub enum SeekFrom {
     Current(i16),
 }
 
+// TODO: use wrapping_add_signed when stabilized
+// https://github.com/rust-lang/rust/issues/87840
+// https://github.com/rust-lang/rust/blob/21b0325c68421b00c6c91055ac330bd5ffe1ea6b/library/core/src/num/uint_macros.rs#L1205
+fn wrapping_add_signed(ptr: u16, offset: i16) -> u16 {
+    ptr.wrapping_add(offset as u16)
+}
+
 impl SeekFrom {
     #[must_use]
     #[inline]
@@ -255,195 +231,55 @@ impl SeekFrom {
     }
 }
 
-// TODO: use wrapping_add_signed when stabilized
-// https://github.com/rust-lang/rust/issues/87840
-// https://github.com/rust-lang/rust/blob/21b0325c68421b00c6c91055ac330bd5ffe1ea6b/library/core/src/num/uint_macros.rs#L1205
-fn wrapping_add_signed(ptr: u16, offset: i16) -> u16 {
-    ptr.wrapping_add(offset as u16)
-}
+/// The `Seek` trait provides a cursor which can be moved within a stream of
+/// bytes.
+///
+/// This is used for navigating the socket buffers, and it is designed to be
+/// similar to [`std::io::Seek`].
+///
+/// [`std::io::Seek`]: https://doc.rust-lang.org/stable/std/io/trait.Seek.html
+pub trait Seek {
+    /// Seek to an offset, in bytes, within the socket buffer.
+    ///
+    /// Seeking beyond the limits will result in wrapping around to the next
+    /// valid address.
+    ///
+    /// # Limits
+    ///
+    /// * [`TcpWriter`] and [`UdpWriter`] are limited by socket free size.
+    /// * [`UdpReader`] is limited by the received size or the UDP datagram length,
+    ///   whichever is less.
+    /// * [`TcpReader`] is limited by the received size.
+    fn seek(&mut self, pos: SeekFrom);
 
-/// Streaming writer for a UDP socket buffer.
-///
-/// Created with [`Udp::udp_writer`].
-///
-/// # Example
-///
-/// ```no_run
-/// # use embedded_hal_mock as h;
-/// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-/// use w5500_hl::{
-///     ll::{Registers, Sn::Sn0},
-///     net::{Ipv4Addr, SocketAddrV4},
-///     Udp,
-///     UdpWriter,
-/// };
-///
-/// const DEST: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 0, 2, 1), 8081);
-///
-/// w5500.udp_bind(Sn0, 8080)?;
-///
-/// let mut udp_writer: UdpWriter<_> = w5500.udp_writer(Sn0)?;
-///
-/// let data_header: [u8; 10] = [0; 10];
-/// let n_written: usize = udp_writer.write(&data_header)?;
-/// assert_eq!(n_written, data_header.len());
-///
-/// let data: [u8; 123] = [0; 123];
-/// let n_written: usize = udp_writer.write(&data)?;
-/// assert_eq!(n_written, data.len());
-///
-/// udp_writer.send_to(&DEST)?;
-/// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-/// ```
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct UdpWriter<'a, W: Registers> {
-    w5500: &'a mut W,
-    sn: Sn,
-    head_ptr: u16,
-    tail_ptr: u16,
-    ptr: u16,
-}
-
-impl<'a, W: Registers> UdpWriter<'a, W> {
-    /// Seek to an offset of the socket buffer.
-    pub fn seek(&mut self, seek: SeekFrom) {
-        self.ptr = seek.new_ptr(self.ptr, self.head_ptr, self.tail_ptr);
-    }
-
-    /// Rewind to the beginning of the socket buffer.
+    /// Rewind to the beginning of the stream.
     ///
     /// This is a convenience method, equivalent to `seek(SeekFrom::Start(0))`.
-    pub fn rewind(&mut self) {
+    fn rewind(&mut self) {
         self.seek(SeekFrom::Start(0))
     }
 
-    /// Returns the current seek position from the start of the socket buffer.
-    #[inline]
-    pub fn position(&self) -> u16 {
-        self.ptr.wrapping_sub(self.head_ptr)
-    }
+    /// Return the length of the stream, in bytes.
+    ///
+    /// * For [`TcpWriter`] and [`UdpWriter`] this returns the socket free size.
+    /// * For [`TcpReader`] this returns the received size.
+    /// * For [`UdpReader`] this returns the received size or the UDP datagram
+    ///   length, whichever is less.
+    fn stream_len(&self) -> u16;
+
+    /// Returns the current seek position from the start of the stream.
+    fn stream_position(&self) -> u16;
 
     /// Remaining bytes in the socket buffer from the current seek position.
-    #[inline]
-    pub fn remain(&self) -> u16 {
-        self.tail_ptr.wrapping_sub(self.ptr)
-    }
-
-    /// Write data to the UDP socket, and return the number of bytes written.
-    ///
-    /// This will return [`nb::Error::WouldBlock`] without modifying the W5500
-    /// socket buffer if there is not enough free space to write all of `buf`.
-    pub fn write(&mut self, buf: &[u8]) -> Result<usize, W::Error> {
-        let write_size: u16 = min(self.remain(), buf.len().try_into().unwrap_or(u16::MAX));
-        if write_size != 0 {
-            self.w5500
-                .set_sn_tx_buf(self.sn, self.ptr, &buf[..(write_size as usize)])?;
-            self.ptr = self.ptr.wrapping_add(write_size);
-
-            Ok(write_size as usize)
-        } else {
-            Ok(0)
-        }
-    }
-
-    /// Writes all the data, returning [`Error::OutOfMemory`] if the size of
-    /// `buf` exceeds the free memory available in the socket buffer.
-    ///
-    /// # Errors
-    ///
-    /// This method can only return:
-    ///
-    /// * [`Error::Other`]
-    /// * [`Error::OutOfMemory`]
-    pub fn write_all(&mut self, buf: &[u8]) -> Result<(), Error<W::Error>> {
-        let buf_len: u16 = buf.len().try_into().unwrap_or(u16::MAX);
-        let write_size: u16 = min(self.remain(), buf_len);
-        if write_size != buf_len {
-            Err(Error::OutOfMemory)
-        } else {
-            self.w5500.set_sn_tx_buf(self.sn, self.ptr, buf)?;
-            self.ptr = self.ptr.wrapping_add(write_size);
-            Ok(())
-        }
-    }
-
-    /// Send all data previously written with [`write`].
-    ///
-    /// The destination is set by the last call to [`Registers::set_sn_dest`],
-    /// [`Udp::send_to`], or [`UdpWrite::send_to`].
-    ///
-    /// [`Registers::set_sn_dest`]: w5500_ll::Registers::set_sn_dest
-    /// [`Udp::send_to`]: Udp::udp_send_to
-    /// [`UdpWrite::send_to`]: UdpWrite::udp_send_to
-    pub fn send(self) -> Result<&'a mut W, W::Error> {
-        self.w5500.set_sn_tx_wr(self.sn, self.ptr)?;
-        self.w5500.set_sn_cr(self.sn, SocketCommand::Send)?;
-        Ok(self.w5500)
-    }
-
-    /// Send all data previously written with [`write`] to the given address.
-    pub fn send_to(self, addr: &SocketAddrV4) -> Result<&'a mut W, W::Error> {
-        self.w5500.set_sn_dest(self.sn, addr)?;
-        self.send()
-    }
+    fn remain(&self) -> u16;
 }
 
-#[derive(Debug)]
-#[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct UdpReader<'a, W: Registers> {
-    w5500: &'a mut W,
-    sn: Sn,
-    header: UdpHeader,
-    head_ptr: u16,
-    tail_ptr: u16,
-    ptr: u16,
-}
-
-impl<'a, W: Registers> UdpReader<'a, W> {
-    /// Get the UDP header.
-    #[inline]
-    pub fn header(&self) -> &UdpHeader {
-        &self.header
-    }
-
-    /// Seek to an offset of the UDP packet.
-    pub fn seek(&mut self, seek: SeekFrom) {
-        self.ptr = seek.new_ptr(self.ptr, self.head_ptr, self.tail_ptr);
-    }
-
-    /// Rewind to the beginning of the UDP packet.
-    ///
-    /// This is a convenience method, equivalent to `seek(SeekFrom::Start(0))`.
-    pub fn rewind(&mut self) {
-        self.seek(SeekFrom::Start(0))
-    }
-
-    /// Returns the current seek position from the start of the UDP packet.
-    #[inline]
-    pub fn position(&self) -> u16 {
-        self.ptr.wrapping_sub(self.head_ptr)
-    }
-
-    /// Remaining bytes in the UDP packet from the current seek position.
-    #[inline]
-    pub fn remain(&self) -> u16 {
-        self.tail_ptr.wrapping_sub(self.ptr)
-    }
-
+/// Socket reader trait.
+///
+/// This is implemented by [`TcpReader`] and [`UdpReader`].
+pub trait Read<'a, W: Registers> {
     /// Read data from the UDP socket, and return the number of bytes read.
-    pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, W::Error> {
-        let read_size: u16 = min(self.remain(), buf.len().try_into().unwrap_or(u16::MAX));
-        if read_size != 0 {
-            self.w5500
-                .sn_rx_buf(self.sn, self.ptr, &mut buf[..(read_size as usize)])?;
-            self.ptr = self.ptr.wrapping_add(read_size);
-
-            Ok(read_size as usize)
-        } else {
-            Ok(0)
-        }
-    }
+    fn read(&mut self, buf: &mut [u8]) -> Result<u16, W::Error>;
 
     /// Read the exact number of bytes required to fill `buf`.
     ///
@@ -456,971 +292,43 @@ impl<'a, W: Registers> UdpReader<'a, W> {
     ///
     /// * [`Error::Other`]
     /// * [`Error::UnexpectedEof`]
-    pub fn read_exact(&mut self, buf: &mut [u8]) -> Result<(), Error<W::Error>> {
-        let buf_len: u16 = buf.len().try_into().unwrap_or(u16::MAX);
-        let read_size: u16 = min(self.remain(), buf_len);
-        if read_size != buf_len {
-            Err(Error::UnexpectedEof)
-        } else {
-            self.w5500.sn_rx_buf(self.sn, self.ptr, buf)?;
-            self.ptr = self.ptr.wrapping_add(read_size);
-            Ok(())
-        }
-    }
+    fn read_exact(&mut self, buf: &mut [u8]) -> Result<(), Error<W::Error>>;
 
-    /// Mark the UDP packet as read, discarding any unread data.
-    pub fn done(self) -> Result<&'a mut W, W::Error> {
-        self.w5500.set_sn_rx_rd(self.sn, self.tail_ptr)?;
-        self.w5500.set_sn_cr(self.sn, SocketCommand::Recv)?;
-        Ok(self.w5500)
-    }
+    /// Mark the data as read, removing the data from the queue.
+    ///
+    /// For a UDP reader this removes the UDP datagram from the queue.
+    fn done(self) -> Result<&'a mut W, W::Error>;
 
-    /// Ignore all read data, and put the UDP packet back into the queue.
-    #[inline]
-    pub fn ignore(self) -> &'a mut W {
-        self.w5500
-    }
+    /// All data and return it to the queue.
+    ///
+    /// For a UDP reader this returns the UDP datagram to the queue.
+    fn ignore(self) -> &'a mut W;
 }
 
-/// A W5500 UDP socket trait.
+/// Socket writer trait.
 ///
-/// After creating a `UdpSocket` by [`bind`]ing it to a socket address,
-/// data can be [sent to] and [received from] any other socket address.
-///
-/// As stated in the User Datagram Protocol's specification in [IETF RFC 768],
-/// UDP is an unordered, unreliable protocol; refer to [`Tcp`] for the TCP trait.
-///
-/// # Comparison to [`std::net::UdpSocket`]
-///
-/// * Everything is non-blocking.
-/// * There is no socket struct, you must pass a socket number as the first
-///   argument to the methods.  This was simply the cleanest solution to the
-///   ownership problem after some experimentation; though it certainly is not
-///   the safest.
-///
-/// [`bind`]: Udp::udp_bind
-/// [IETF RFC 768]: https://tools.ietf.org/html/rfc768
-/// [received from]: Udp::udp_recv_from
-/// [sent to]: Udp::udp_send_to
-/// [`Tcp`]: crate::Tcp
-/// [`std::net::UdpSocket`]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html
-pub trait Udp: Registers {
-    /// Binds the socket to the given port.
-    ///
-    /// This will close the socket, which will reset the RX and TX buffers.
-    ///
-    /// # Comparison to [`std::net::UdpSocket::bind`]
-    ///
-    /// This method accepts a port instead of a [`net::SocketAddrV4`], this is
-    /// because the IP address is global for the device, set by the
-    /// [source IP register], and cannot be set on a per-socket basis.
-    ///
-    /// Additionally you can only provide one port, instead of iterable
-    /// addresses to bind.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The port must not be in use by any other socket on the W5500.
-    ///
-    /// # Example
-    ///
-    /// Bind the first socket to port 8080.
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::ll::{Registers, Sn::Sn0};
-    /// use w5500_hl::Udp;
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`std::net::UdpSocket::bind`]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.bind
-    /// [source IP register]: w5500_ll::Registers::sipr
-    fn udp_bind(&mut self, sn: Sn, port: u16) -> Result<(), Self::Error> {
-        debug_assert!(
-            port_is_unique(self, sn, port)?,
-            "Local port {} is in use",
-            port
-        );
+/// This is implemented by [`TcpWriter`] and [`UdpWriter`].
+pub trait Write<'a, W: Registers> {
+    /// Write data to the socket buffer, and return the number of bytes written.
+    fn write(&mut self, buf: &[u8]) -> Result<u16, W::Error>;
 
-        self.set_sn_cr(sn, SocketCommand::Close)?;
-        // This will not hang, the socket status will always change to closed
-        // after a close command.
-        // (unless you do somthing silly like holding the W5500 in reset)
-        loop {
-            if self.sn_sr(sn)? == Ok(SocketStatus::Closed) {
-                break;
-            }
-        }
-        self.set_sn_port(sn, port)?;
-        const MODE: SocketMode = SocketMode::DEFAULT.set_protocol(Protocol::Udp);
-        self.set_sn_mr(sn, MODE)?;
-        self.set_sn_cr(sn, SocketCommand::Open)?;
-        // This will not hang, the socket status will always change to Udp
-        // after a open command with SN_MR set to UDP.
-        // (unless you do somthing silly like holding the W5500 in reset)
-        loop {
-            if self.sn_sr(sn)? == Ok(SocketStatus::Udp) {
-                break;
-            }
-        }
-        Ok(())
-    }
-
-    /// Receives a single datagram message on the socket.
-    /// On success, returns the number of bytes read and the origin.
-    ///
-    /// The function must be called with valid byte array `buf` of sufficient
-    /// size to hold the message bytes.
-    /// If a message is too long to fit in the supplied buffer, excess bytes
-    /// will be discarded.
-    ///
-    /// # Comparison to [`std::net::UdpSocket::recv_from`]
-    ///
-    /// * This method will always discard excess bytes from the socket buffer.
-    /// * This method is non-blocking, use [`block`] to treat it as blocking.
+    /// Writes all the data, returning [`Error::OutOfMemory`] if the size of
+    /// `buf` exceeds the free memory available in the socket buffer.
     ///
     /// # Errors
     ///
     /// This method can only return:
     ///
     /// * [`Error::Other`]
-    /// * [`Error::WouldBlock`]
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     block,
-    ///     Udp,
-    /// };
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let mut buf = [0; 10];
-    /// let (number_of_bytes, src_addr) = block!(w5500.udp_recv_from(Sn0, &mut buf))?;
-    ///
-    /// // panics if bytes were discarded
-    /// assert!(
-    ///     number_of_bytes < buf.len(),
-    ///     "Buffer was too small to receive all data"
-    /// );
-    ///
-    /// let filled_buf = &mut buf[..number_of_bytes];
-    /// # Ok::<(), w5500_hl::Error<_>>(())
-    /// ```
-    ///
-    /// [`std::net::UdpSocket::recv_from`]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.recv_from
-    fn udp_recv_from(
-        &mut self,
-        sn: Sn,
-        buf: &mut [u8],
-    ) -> Result<(usize, SocketAddrV4), Error<Self::Error>> {
-        let rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UdpHeader::LEN) {
-            Some(rsr) => rsr,
-            // nothing to recieve
-            None => return Err(Error::WouldBlock),
-        };
+    /// * [`Error::OutOfMemory`]
+    fn write_all(&mut self, buf: &[u8]) -> Result<(), Error<W::Error>>;
 
-        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
-
-        let mut ptr: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
-        self.sn_rx_buf(sn, ptr, &mut header)?;
-        ptr = ptr.wrapping_add(UdpHeader::LEN);
-        let header: UdpHeader = UdpHeader::deser(header);
-
-        // not all data as indicated by the header has been buffered
-        if rsr < header.len {
-            return Err(Error::WouldBlock);
-        }
-
-        let read_size: usize = min(usize::from(header.len), buf.len());
-        if read_size != 0 {
-            self.sn_rx_buf(sn, ptr, &mut buf[..read_size])?;
-        }
-        ptr = ptr.wrapping_add(header.len);
-        self.set_sn_rx_rd(sn, ptr)?;
-        self.set_sn_cr(sn, SocketCommand::Recv)?;
-        Ok((read_size, header.origin))
-    }
-
-    /// Receives a single datagram message on the socket, without removing it
-    /// from the queue.
-    /// On success, returns the number of bytes read and the UDP header.
-    ///
-    /// # Comparison to [`std::net::UdpSocket::peek_from`]
-    ///
-    /// * This method will never discard excess bytes from the socket buffer.
-    /// * This method is non-blocking, use [`block`] to treat it as blocking.
-    ///
-    /// # Errors
-    ///
-    /// This method can only return:
-    ///
-    /// * [`Error::Other`]
-    /// * [`Error::WouldBlock`]
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     block,
-    ///     Udp,
-    /// };
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let mut buf = [0; 10];
-    /// let (number_of_bytes, udp_header) = block!(w5500.udp_peek_from(Sn0, &mut buf))?;
-    ///
-    /// // panics if buffer was too small
-    /// assert!(
-    ///     number_of_bytes > buf.len(),
-    ///     "Buffer was of len {} too small to receive all data: {} / {} bytes read",
-    ///     buf.len(), number_of_bytes, udp_header.len
-    /// );
-    ///
-    /// let filled_buf = &mut buf[..number_of_bytes];
-    /// # Ok::<(), w5500_hl::Error<_>>(())
-    /// ```
-    ///
-    /// [`std::net::UdpSocket::peek_from`]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.peek_from
-    fn udp_peek_from(
-        &mut self,
-        sn: Sn,
-        buf: &mut [u8],
-    ) -> Result<(usize, UdpHeader), Error<Self::Error>> {
-        let rsr: u16 = match self.sn_rx_rsr(sn)?.checked_sub(UdpHeader::LEN) {
-            Some(rsr) => rsr,
-            // nothing to recieve
-            None => return Err(Error::WouldBlock),
-        };
-
-        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
-
-        let mut ptr: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
-        self.sn_rx_buf(sn, ptr, &mut header)?;
-        ptr = ptr.wrapping_add(UdpHeader::LEN);
-        let header: UdpHeader = UdpHeader::deser(header);
-
-        // not all data as indicated by the header has been buffered
-        if rsr < header.len {
-            return Err(Error::WouldBlock);
-        }
-
-        let read_size: usize = min(usize::from(header.len), buf.len());
-        if read_size != 0 {
-            self.sn_rx_buf(sn, ptr, &mut buf[..read_size])?;
-        }
-
-        Ok((read_size, header))
-    }
-
-    /// Receives the origin and size of the next datagram avaliable on the
-    /// socket, without removing it from the queue.
-    ///
-    /// There is no [`std::net`](https://doc.rust-lang.org/std/net) equivalent
-    /// for this method.
-    ///
-    /// # Errors
-    ///
-    /// This method can only return:
-    ///
-    /// * [`Error::Other`]
-    /// * [`Error::WouldBlock`]
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     Udp, UdpHeader, block
-    /// };
-    /// // global_allocator is currently avaliable on nightly for embedded rust
-    /// extern crate alloc;
-    /// use alloc::vec::{self, Vec};
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let udp_header: UdpHeader = block!(w5500.udp_peek_from_header(Sn0))?;
-    /// let bytes_to_allocate: usize = udp_header.len as usize;
-    ///
-    /// let mut buf: Vec<u8> = vec![0; bytes_to_allocate];
-    /// let (number_of_bytes, source) = block!(w5500.udp_recv_from(Sn0, &mut buf))?;
-    /// debug_assert_eq!(bytes_to_allocate, number_of_bytes);
-    /// # Ok::<(), w5500_hl::Error<_>>(())
-    /// ```
-    fn udp_peek_from_header(&mut self, sn: Sn) -> Result<UdpHeader, Error<Self::Error>> {
-        let rsr: u16 = self.sn_rx_rsr(sn)?;
-
-        // nothing to recieve
-        if rsr < UdpHeader::LEN {
-            return Err(Error::WouldBlock);
-        }
-
-        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
-
-        let ptr: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
-        self.sn_rx_buf(sn, ptr, &mut header)?;
-        Ok(UdpHeader::deser(header))
-    }
-
-    /// Sends data on the socket to the given address.
-    /// On success, returns the number of bytes written.
-    ///
-    /// # Comparison to [`std::net::UdpSocket::send_to`]
-    ///
-    /// * You cannot transmit more than `u16::MAX` bytes at once.
-    /// * You can only provide one destination.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Udp,
-    /// };
-    ///
-    /// const DEST: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 0, 2, 1), 8081);
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let buf: [u8; 10] = [0; 10];
-    /// let tx_bytes = w5500.udp_send_to(Sn0, &buf, &DEST)?;
-    /// assert_eq!(tx_bytes, buf.len());
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`std::net::UdpSocket::send_to`]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.send_to
-    fn udp_send_to(
-        &mut self,
-        sn: Sn,
-        buf: &[u8],
-        addr: &SocketAddrV4,
-    ) -> Result<usize, Self::Error> {
-        self.set_sn_dest(sn, addr)?;
-        self.udp_send(sn, buf)
-    }
-
-    /// Sends data on the socket to the given address.
-    /// On success, returns the number of bytes written.
-    ///
-    /// This will transmit only if there is enough free space in the W5500
-    /// transmit buffer.
-    ///
-    /// # Comparison to [`std::net::UdpSocket::send_to`]
-    ///
-    /// * You cannot transmit more than `u16::MAX` bytes at once.
-    /// * You can only provide one destination.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Udp,
-    /// };
-    ///
-    /// const DEST: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 0, 2, 1), 8081);
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let buf: [u8; 10] = [0; 10];
-    /// let tx_bytes = w5500.udp_send_to_if_free(Sn0, &buf, &DEST)?;
-    /// assert_eq!(tx_bytes, buf.len());
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`std::net::UdpSocket::send_to`]: https://doc.rust-lang.org/std/net/struct.UdpSocket.html#method.send_to
-    fn udp_send_to_if_free(
-        &mut self,
-        sn: Sn,
-        buf: &[u8],
-        addr: &SocketAddrV4,
-    ) -> Result<usize, Self::Error> {
-        self.set_sn_dest(sn, addr)?;
-        self.udp_send_if_free(sn, buf)
-    }
-
-    /// Sends data to the currently configured destination.
-    /// On success, returns the number of bytes written.
-    ///
-    /// The destination is set by the last call to [`Registers::set_sn_dest`],
-    /// [`Udp::send_to`], or [`UdpWrite::send_to`].
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Udp,
-    /// };
-    ///
-    /// const DEST: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 0, 2, 1), 8081);
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let buf: [u8; 10] = [0; 10];
-    /// let tx_bytes: usize = w5500.udp_send_to(Sn0, &buf, &DEST)?;
-    /// assert_eq!(tx_bytes, buf.len());
-    /// // send the same to the same destination
-    /// let tx_bytes: usize = w5500.udp_send(Sn0, &buf)?;
-    /// assert_eq!(tx_bytes, buf.len());
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`Registers::set_sn_dest`]: w5500_ll::Registers::set_sn_dest
-    /// [`Udp::send_to`]: Udp::udp_send_to
-    /// [`UdpWrite::send_to`]: UdpWrite::udp_send_to
-    fn udp_send(&mut self, sn: Sn, buf: &[u8]) -> Result<usize, Self::Error> {
-        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
-
-        let data_len: u16 = u16::try_from(buf.len()).unwrap_or(u16::MAX);
-        let free_size: u16 = self.sn_tx_fsr(sn)?;
-        let tx_bytes: u16 = min(data_len, free_size);
-        if tx_bytes != 0 {
-            let ptr: u16 = self.sn_tx_wr(sn)?;
-            self.set_sn_tx_buf(sn, ptr, &buf[..usize::from(tx_bytes)])?;
-            self.set_sn_tx_wr(sn, ptr.wrapping_add(tx_bytes))?;
-            self.set_sn_cr(sn, SocketCommand::Send)?;
-        }
-        Ok(usize::from(tx_bytes))
-    }
-
-    /// Sends data to the currently configured destination.
-    /// On success, returns the number of bytes written.
-    ///
-    /// The destination is set by the last call to [`Registers::set_sn_dest`],
-    /// [`Udp::send_to`], or [`UdpWrite::send_to`].
-    ///
-    /// This will transmit only if there is enough free space in the W5500
-    /// transmit buffer.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be opened as a UDP socket.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn::Sn0},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Udp,
-    /// };
-    ///
-    /// const DEST: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 0, 2, 1), 8081);
-    ///
-    /// w5500.udp_bind(Sn0, 8080)?;
-    /// let buf: [u8; 10] = [0; 10];
-    /// let tx_bytes: usize = w5500.udp_send_to_if_free(Sn0, &buf, &DEST)?;
-    /// assert_eq!(tx_bytes, buf.len());
-    /// // send the same to the same destination
-    /// let tx_bytes: usize = w5500.udp_send_if_free(Sn0, &buf)?;
-    /// assert_eq!(tx_bytes, buf.len());
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`Registers::set_sn_dest`]: w5500_ll::Registers::set_sn_dest
-    /// [`Udp::send_to`]: Udp::udp_send_to
-    /// [`UdpWrite::send_to`]: UdpWrite::udp_send_to
-    fn udp_send_if_free(&mut self, sn: Sn, buf: &[u8]) -> Result<usize, Self::Error> {
-        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
-
-        let data_len: u16 = match u16::try_from(buf.len()) {
-            Ok(l) => l,
-            Err(_) => return Ok(0),
-        };
-        let free_size: u16 = self.sn_tx_fsr(sn)?;
-        if data_len <= free_size {
-            let ptr: u16 = self.sn_tx_wr(sn)?;
-            self.set_sn_tx_buf(sn, ptr, buf)?;
-            self.set_sn_tx_wr(sn, ptr.wrapping_add(data_len))?;
-            self.set_sn_cr(sn, SocketCommand::Send)?;
-        }
-        Ok(usize::from(data_len))
-    }
-
-    /// Create a UDP writer.
-    ///
-    /// This returns a [`UdpWriter`] structure, which contains functions to
-    /// stream data into the W5500 socket buffers incrementally.
-    ///
-    /// # Example
-    ///
-    /// See [`UdpWriter`].
-    fn udp_writer(&mut self, sn: Sn) -> Result<UdpWriter<Self>, Self::Error>
-    where
-        Self: Sized,
-    {
-        let sn_tx_fsr: u16 = self.sn_tx_fsr(sn)?;
-        let sn_tx_wr: u16 = self.sn_tx_wr(sn)?;
-
-        Ok(UdpWriter {
-            w5500: self,
-            sn,
-            head_ptr: sn_tx_wr,
-            tail_ptr: sn_tx_wr.wrapping_add(sn_tx_fsr),
-            ptr: sn_tx_wr,
-        })
-    }
-
-    /// Create a UDP reader.
-    ///
-    /// This returns a [`UdpReader`] structure, which contains functions to
-    /// stream data from the W5500 socket buffers incrementally.
-    ///
-    /// This will return [`nb::Result::WouldBlock`] if there is no data to read.
-    ///
-    /// # Errors
-    ///
-    /// This method can only return:
-    ///
-    /// * [`Error::Other`]
-    /// * [`Error::WouldBlock`]
-    ///
-    /// # Example
-    ///
-    /// See [`UdpReader`].
-    fn udp_reader(&mut self, sn: Sn) -> Result<UdpReader<Self>, Error<Self::Error>>
-    where
-        Self: Sized,
-    {
-        let sn_rx_rsr: u16 = self.sn_rx_rsr(sn)?;
-        if sn_rx_rsr < UdpHeader::LEN {
-            return Err(Error::WouldBlock);
-        }
-
-        debug_assert_eq!(self.sn_sr(sn)?, Ok(SocketStatus::Udp));
-
-        let sn_rx_rd: u16 = self.sn_rx_rd(sn)?;
-        let mut header: [u8; UdpHeader::LEN_USIZE] = [0; UdpHeader::LEN_USIZE];
-        self.sn_rx_buf(sn, sn_rx_rd, &mut header)?;
-        let header: UdpHeader = UdpHeader::deser(header);
-
-        let head_ptr: u16 = sn_rx_rd.wrapping_add(UdpHeader::LEN);
-
-        Ok(UdpReader {
-            w5500: self,
-            sn,
-            head_ptr,
-            tail_ptr: sn_rx_rd.wrapping_add(sn_rx_rsr),
-            ptr: head_ptr,
-            header,
-        })
-    }
+    /// Send all data previously written with [`write`] and [`write_all`].
+    ///
+    /// For [`UdpWriter`] the destination is set by the last call to
+    /// [`Registers::set_sn_dest`], [`Udp::send_to`], or [`UdpWriter::send_to`].
+    fn send(self) -> Result<&'a mut W, W::Error>;
 }
-
-/// Implement the UDP trait for any structure that implements [`w5500_ll::Registers`].
-impl<T> Udp for T where T: Registers {}
-
-/// A W5500 TCP trait.
-pub trait Tcp: Registers {
-    /// Starts the 3-way TCP handshake with the remote host.
-    ///
-    /// This method is used to create and interact with a TCP stream between
-    /// a local host and a remote socket.
-    ///
-    /// After initiating a connection with [`tcp_connect`] and recieving the
-    /// [`con`] interrupt data can be transmitting by using [`tcp_read`] and
-    /// [`tcp_write`].
-    ///
-    /// Calling this method **does not** mean the socket will be connected
-    /// afterwards, this simply starts the three way handshake.
-    ///
-    /// After calling this method you will eventually get one of 3 interrupts on
-    /// the socket:
-    ///
-    /// 1. [`con`](w5500_ll::SocketInterrupt::con_raised)
-    /// 2. [`discon`](w5500_ll::SocketInterrupt::discon_raised)
-    /// 3. [`timeout`](w5500_ll::SocketInterrupt::timeout_raised)
-    ///
-    /// # Arguments
-    ///
-    /// * `socket` - The socket number to use for this TCP stream.
-    /// * `port` - The local port to use for the TCP connection.
-    /// * `addr` - Address of the remote host to connect to.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The port must not be in use by any other socket on the W5500.
-    ///
-    /// # Example
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn, SocketInterrupt},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Tcp,
-    /// };
-    ///
-    /// const MQTT_SOCKET: Sn = Sn::Sn0;
-    /// const MQTT_SOURCE_PORT: u16 = 33650;
-    /// const MQTT_SERVER: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 168, 2, 10), 1883);
-    ///
-    /// w5500.tcp_connect(MQTT_SOCKET, MQTT_SOURCE_PORT, &MQTT_SERVER)?;
-    ///
-    /// // wait for a socket interrupt
-    /// // you should use the actual interrupt pin, polling is just for demonstration
-    /// loop {
-    ///     let sn_ir: SocketInterrupt = w5500.sn_ir(MQTT_SOCKET)?;
-    ///
-    ///     // in reality you will want to handle disconnections gracefully with retries
-    ///     assert!(!sn_ir.discon_raised());
-    ///     assert!(!sn_ir.timeout_raised());
-    ///
-    ///     // connection succeded
-    ///     if sn_ir.con_raised() {
-    ///         break;
-    ///     }
-    /// }
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`tcp_write`]: Tcp::tcp_write
-    /// [`tcp_read`]: Tcp::tcp_read
-    /// [`tcp_connect`]: Tcp::tcp_connect
-    /// [`con`]: w5500_ll::SocketInterrupt::con_raised
-    fn tcp_connect(&mut self, sn: Sn, port: u16, addr: &SocketAddrV4) -> Result<(), Self::Error> {
-        debug_assert!(
-            port_is_unique(self, sn, port)?,
-            "Local port {} is in use",
-            port
-        );
-
-        self.set_sn_cr(sn, SocketCommand::Close)?;
-        // This will not hang, the socket status will always change to closed
-        // after a close command.
-        // (unless you do somthing silly like holding the W5500 in reset)
-        loop {
-            if self.sn_sr(sn)? == Ok(SocketStatus::Closed) {
-                break;
-            }
-        }
-        const MODE: SocketMode = SocketMode::DEFAULT.set_protocol(Protocol::Tcp);
-        self.set_sn_mr(sn, MODE)?;
-        self.set_sn_port(sn, port)?;
-        self.set_sn_cr(sn, SocketCommand::Open)?;
-        self.set_sn_dest(sn, addr)?;
-        // This will not hang, the socket status will always change to Init
-        // after a open command with SN_MR set to TCP.
-        // (unless you do somthing silly like holding the W5500 in reset)
-        loop {
-            if self.sn_sr(sn)? == Ok(SocketStatus::Init) {
-                break;
-            }
-        }
-        self.set_sn_cr(sn, SocketCommand::Connect)
-    }
-
-    /// Open a TCP listener on the given port.
-    ///
-    /// After opening a listener with [`tcp_listen`] and recieving the
-    /// [`con`] interrupt data can be transmitting by using [`tcp_read`] and
-    /// [`tcp_write`].
-    ///
-    /// # Arguments
-    ///
-    /// * `socket` - The socket number to use for this TCP listener.
-    /// * `port` - The local port to listen for remote connections on.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The port must not be in use by any other socket on the W5500.
-    ///
-    /// # Example
-    ///
-    /// Create an HTTP server.
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn, SocketInterrupt},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Tcp,
-    /// };
-    /// // global_allocator is currently avaliable on nightly for embedded rust
-    /// extern crate alloc;
-    /// use alloc::vec::{self, Vec};
-    ///
-    /// const HTTP_SOCKET: Sn = Sn::Sn1;
-    /// const HTTP_PORT: u16 = 80;
-    ///
-    /// // start serving
-    /// w5500.tcp_listen(HTTP_SOCKET, HTTP_PORT)?;
-    ///
-    /// // wait for the RECV interrupt, indicating there is data to read from a client
-    /// loop {
-    ///     let sn_ir = w5500.sn_ir(HTTP_SOCKET).unwrap();
-    ///     if sn_ir.recv_raised() {
-    ///         w5500.set_sn_ir(HTTP_SOCKET, sn_ir).unwrap();
-    ///         break;
-    ///     }
-    ///     if sn_ir.discon_raised() | sn_ir.timeout_raised() {
-    ///         panic!("Socket disconnected while waiting for RECV");
-    ///     }
-    /// }
-    ///
-    /// let mut buf: Vec<u8> = vec![0; 256];
-    /// let rx_bytes: usize = w5500.tcp_read(HTTP_SOCKET, &mut buf).unwrap();
-    /// // Truncate the buffer to the number of bytes read
-    /// // Safety: BUF is only borrowed mutably in one location
-    /// let filled_buf: &[u8] = &buf[..rx_bytes];
-    ///
-    /// // parse HTTP request here using filled_buf
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`tcp_write`]: Tcp::tcp_write
-    /// [`tcp_read`]: Tcp::tcp_read
-    /// [`tcp_listen`]: Tcp::tcp_listen
-    /// [`con`]: w5500_ll::SocketInterrupt::con_raised
-    fn tcp_listen(&mut self, sn: Sn, port: u16) -> Result<(), Self::Error> {
-        debug_assert!(
-            port_is_unique(self, sn, port)?,
-            "Local port {} is in use",
-            port
-        );
-
-        self.set_sn_cr(sn, SocketCommand::Close)?;
-        // This will not hang, the socket status will always change to closed
-        // after a close command.
-        // (unless you do somthing silly like holding the W5500 in reset)
-        loop {
-            if self.sn_sr(sn)? == Ok(SocketStatus::Closed) {
-                break;
-            }
-        }
-        const MODE: SocketMode = SocketMode::DEFAULT.set_protocol(Protocol::Tcp);
-        self.set_sn_mr(sn, MODE)?;
-        self.set_sn_port(sn, port)?;
-        self.set_sn_cr(sn, SocketCommand::Open)?;
-        // This will not hang, the socket status will always change to Init
-        // after a open command with SN_MR set to TCP.
-        // (unless you do somthing silly like holding the W5500 in reset)
-        loop {
-            if self.sn_sr(sn)? == Ok(SocketStatus::Init) {
-                break;
-            }
-        }
-        self.set_sn_cr(sn, SocketCommand::Listen)
-    }
-
-    /// Read data from the remote host, returning the number of bytes read.
-    ///
-    /// You should wait for the socket [`recv`] interrupt before calling this method.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be an [`Established`] TCP socket.
-    ///
-    /// # Example
-    ///
-    /// Send a MQTT CONNECT packet and read a CONNACK.
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn, SocketInterrupt},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Tcp,
-    /// };
-    ///
-    /// const MQTT_SOCKET: Sn = Sn::Sn0;
-    /// const MQTT_SOURCE_PORT: u16 = 33650;
-    /// const MQTT_SERVER: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 168, 2, 10), 1883);
-    ///
-    /// w5500.tcp_connect(MQTT_SOCKET, MQTT_SOURCE_PORT, &MQTT_SERVER)?;
-    ///
-    /// // ... wait for a CON interrupt
-    ///
-    /// const CONNECT: [u8; 14] = [
-    ///     0x10, 0x0C, 0x00, 0x04, b'M', b'Q', b'T', b'T', 0x04, 0x02, 0x0E, 0x10, 0x00, 0x00,
-    /// ];
-    /// let tx_bytes: usize = w5500.tcp_write(MQTT_SOCKET, &CONNECT)?;
-    /// assert_eq!(tx_bytes, CONNECT.len());
-    ///
-    /// // ... wait for a RECV interrupt
-    ///
-    /// let mut buf = [0; 10];
-    /// let rx_bytes: usize = w5500.tcp_read(MQTT_SOCKET, &mut buf)?;
-    /// let filled_buf = &buf[..rx_bytes];
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`Established`]: w5500_ll::SocketStatus::Established
-    /// [`recv`]: w5500_ll::SocketInterrupt::recv_raised
-    fn tcp_read(&mut self, sn: Sn, buf: &mut [u8]) -> Result<usize, Self::Error> {
-        debug_assert!(!matches!(
-            self.sn_sr(sn)?,
-            Ok(SocketStatus::Udp) | Ok(SocketStatus::Init) | Ok(SocketStatus::Macraw)
-        ));
-
-        let rx_bytes: u16 = {
-            let rsr: u16 = self.sn_rx_rsr(sn)?;
-            min(rsr, u16::try_from(buf.len()).unwrap_or(u16::MAX))
-        };
-        if rx_bytes != 0 {
-            let ptr: u16 = self.sn_rx_rd(sn)?;
-            self.sn_rx_buf(sn, ptr, &mut buf[..usize::from(rx_bytes)])?;
-            self.set_sn_rx_rd(sn, ptr.wrapping_add(rx_bytes))?;
-            self.set_sn_cr(sn, SocketCommand::Recv)?;
-        }
-        Ok(usize::from(rx_bytes))
-    }
-
-    /// Send data to the remote host, returning the number of bytes written.
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be an [`Established`] TCP socket.
-    ///
-    /// # Example
-    ///
-    /// Send a MQTT CONNECT packet.
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn, SocketInterrupt},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Tcp,
-    /// };
-    ///
-    /// const MQTT_SOCKET: Sn = Sn::Sn0;
-    /// const MQTT_SOURCE_PORT: u16 = 33650;
-    /// const MQTT_SERVER: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 168, 2, 10), 1883);
-    ///
-    /// w5500.tcp_connect(MQTT_SOCKET, MQTT_SOURCE_PORT, &MQTT_SERVER)?;
-    ///
-    /// // ... wait for a CON interrupt
-    ///
-    /// const CONNECT: [u8; 14] = [
-    ///     0x10, 0x0C, 0x00, 0x04, b'M', b'Q', b'T', b'T', 0x04, 0x02, 0x0E, 0x10, 0x00, 0x00,
-    /// ];
-    /// let tx_bytes: usize = w5500.tcp_write(MQTT_SOCKET, &CONNECT)?;
-    /// assert_eq!(tx_bytes, CONNECT.len());
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`Established`]: w5500_ll::SocketStatus::Established
-    fn tcp_write(&mut self, sn: Sn, buf: &[u8]) -> Result<usize, Self::Error> {
-        debug_assert!(!matches!(
-            self.sn_sr(sn)?,
-            Ok(SocketStatus::Udp) | Ok(SocketStatus::Init) | Ok(SocketStatus::Macraw)
-        ));
-
-        let tx_bytes: u16 = {
-            let data_len: u16 = u16::try_from(buf.len()).unwrap_or(u16::MAX);
-            let free_size: u16 = self.sn_tx_fsr(sn)?;
-            min(data_len, free_size)
-        };
-        if tx_bytes != 0 {
-            let ptr: u16 = self.sn_tx_wr(sn)?;
-            self.set_sn_tx_buf(sn, ptr, &buf[..usize::from(tx_bytes)])?;
-            self.set_sn_tx_wr(sn, ptr.wrapping_add(tx_bytes))?;
-            self.set_sn_cr(sn, SocketCommand::Send)?;
-        }
-        Ok(usize::from(tx_bytes))
-    }
-
-    /// Disconnect from the peer.
-    ///
-    /// If the disconnect is successful (FIN/ACK packet is received) the socket
-    /// status changes to [`Closed`], otherwise TCP<sub>TO</sub> occurs, the
-    /// [timeout interrupt] is raised, and the socket status changes to
-    /// [`Closed`].
-    ///
-    /// # Panics
-    ///
-    /// * (debug) The socket must be an [`Established`] TCP socket.
-    ///
-    /// # Example
-    ///
-    /// Connect and disconnect from a MQTT server.
-    ///
-    /// ```no_run
-    /// # use embedded_hal_mock as h;
-    /// # let mut w5500 = w5500_ll::blocking::vdm::W5500::new(h::spi::Mock::new(&[]), h::pin::Mock::new(&[]));
-    /// use w5500_hl::{
-    ///     ll::{Registers, Sn, SocketInterrupt},
-    ///     net::{Ipv4Addr, SocketAddrV4},
-    ///     Tcp,
-    /// };
-    ///
-    /// const MQTT_SOCKET: Sn = Sn::Sn0;
-    /// const MQTT_SOURCE_PORT: u16 = 33650;
-    /// const MQTT_SERVER: SocketAddrV4 = SocketAddrV4::new(Ipv4Addr::new(192, 168, 2, 10), 1883);
-    ///
-    /// w5500.tcp_connect(MQTT_SOCKET, MQTT_SOURCE_PORT, &MQTT_SERVER)?;
-    ///
-    /// // ... wait for a CON interrupt
-    ///
-    /// w5500.tcp_disconnect(MQTT_SOCKET)?;
-    /// # Ok::<(), w5500_hl::ll::blocking::vdm::Error<_, _>>(())
-    /// ```
-    ///
-    /// [`Closed`]: w5500_ll::SocketStatus::Closed
-    /// [`Established`]: w5500_ll::SocketStatus::Established
-    /// [timeout interrupt]: w5500_ll::SocketInterrupt::timeout_raised
-    fn tcp_disconnect(&mut self, sn: Sn) -> Result<(), Self::Error> {
-        debug_assert!(!matches!(
-            self.sn_sr(sn)?,
-            Ok(SocketStatus::Udp) | Ok(SocketStatus::Init) | Ok(SocketStatus::Macraw)
-        ));
-        self.set_sn_cr(sn, SocketCommand::Disconnect)
-    }
-}
-
-/// Implement the TCP trait for any structure that implements [`w5500_ll::Registers`].
-impl<T> Tcp for T where T: Registers {}
 
 /// Methods common to all W5500 socket types.
 pub trait Common: Registers {
